refactor(member): rename api file
