+++ frontend/src/api/index.ts
@@ -1,3 +1,3 @@
-export * from "./member";
-export * from "./recruitments";
 export * from "./applicationForms";
+export * from "./members";
+export * from "./recruitments";
